made min and max integers

+++ Test/rUnif.cpp
@@ -3,7 +3,7 @@
 #include <ctime>
 using namespace std;
 
-double* rUnif(int n, double min = 0, double max = 1)
+double* rUnif(int n, int min = 0, int max = 1)
 {
     if (n < 1)
     {
@@ -19,7 +19,7 @@ double* rUnif(int n, double min = 0, double max = 1)
     static double sample[n];
     for (int i = 0; i < n; i++)
     {
-
+        
     }
 }
